requirements.txt
openai
langchain
langchain_openai
langchain_community
langchain_experimental
sentence-transformers
tf-keras
einops
pypdf
streamlit
tiktoken
python-dotenv
milvus==2.2.16
pymilvus==2.4.1
streamlit-modal
panel
streamlit-option-menu
streamlit-pdf-viewer
openpyxl
matplotlib
seaborn
lark
PyMuPDF
pdfminer.six
pi-heif
pytesseract
unstructured
unstructured.pytesseract
unstructured-inference
langid
tabula
streamlit-feedback
pyperclip